rfc2889/requirements.txt
tdl-xoa-converter>=1.1.0
<<<<<<< zhanisch/dev/rfc2889
tdl-xoa-core>=1.2.0
tdl-xoa-driver>=1.3.0
=======
tdl-xoa-core>=1.1.0
tdl-xoa-driver<1.6
>>>>>>> main
